Single Page for "Manage Account," new redux logic
